+++ pkgs/test/lib/src/runner/browser/chromium.dart
@@ -32,8 +32,8 @@ enum ChromiumBasedBrowser {
     var args = [
       '--user-data-dir=$dir',
       url.toString(),
-      '--enable-logging=stdout',
-      '--v=1',
+      '--enable-logging=stderr',
+      '--v=0',
       '--disable-extensions',
       '--disable-popup-blocking',
       '--bwsi',