+++ pkgs/test/lib/src/runner/browser/platform.dart
@@ -274,13 +274,17 @@ class BrowserPlatform extends PlatformPlugin
 
     if (_closed) return null;
 
-    // TODO(nweiz): Don't start the browser until all the suites are compiled.
     var browserManager = await _browserManagerFor(browser);
     if (_closed || browserManager == null) return null;
 
+    var timeout = const Duration(seconds: 30);
+    if (suiteConfig.metadata.timeout.apply(timeout) case final suiteTimeout?
+        when suiteTimeout > timeout) {
+      timeout = suiteTimeout;
+    }
     var suite = await browserManager.load(
         path, suiteUrl, suiteConfig, message, platform.compiler,
-        mapper: _mappers[path]);
+        mapper: _mappers[path], timeout: timeout);
     if (_closed) return null;
     return suite;
   }