+++ pkgs/test/test/runner/browser/internet_explorer_test.dart
@@ -43,7 +43,7 @@ webSocket.addEventListener("open", function() {
   });
 
   test('reports an error in onExit', () {
-    var ie = InternetExplorer('http://dart-lang.org',
+    var ie = InternetExplorer(Uri.https('dart.dev'),
         settings: ExecutableSettings(
             linuxExecutable: '_does_not_exist',
             macOSExecutable: '_does_not_exist',