+++ pkgs/test_core/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.5.7
+
+* Pass --disable-program-split to dart2js to fix tests which use deferred
+  loading.
+* Add a 'silent' reporter option. Keep it hidden in the CLI args help since it
+  is not useful in the general case, but can be useful for tests of the test
+  runner.
+* Update to `package:vm_service` `12.0.0`
+
 ## 0.5.6
 
 * Add support for discontinuing after the first failing test with `--fail-fast`.